+++ src/contentScript/messenger/registration.ts
@@ -32,9 +32,6 @@ import {
   removePersistedExtension,
 } from "@/contentScript/lifecycle"; // 202 strictNullCheck errors
 import {
-  clearDynamicElements,
-  disableOverlay,
-  enableOverlay,
   runExtensionPointReader,
   updateDynamicElement,
 } from "@/contentScript/pageEditor/dynamic"; // 205 strictNullCheck errors
@@ -51,6 +48,7 @@ import {
 } from "@/contentScript/pipelineProtocol"; // Depends on background/messenger to pass strictNullCheck
 import { getCopilotHostData } from "@/contrib/automationanywhere/SetCopilotDataEffect"; // Depends on background/messenger to pass strictNullCheck
 import { showBannerFromConfig } from "@/contentScript/integrations/deferredLoginController"; // Depends on background/messenger to pass strictNullCheck
+import { clearDynamicElements } from "@/contentScript/pageEditor/dynamic/clearDynamicElements"; // 201 strictNullCheck errors
 
 expectContext("contentScript");
 
@@ -67,8 +65,7 @@ declare global {
     CLEAR_DYNAMIC_ELEMENTS: typeof clearDynamicElements;
     UPDATE_DYNAMIC_ELEMENT: typeof updateDynamicElement;
     RUN_EXTENSION_POINT_READER: typeof runExtensionPointReader;
-    ENABLE_OVERLAY: typeof enableOverlay;
-    DISABLE_OVERLAY: typeof disableOverlay;
+
     INSTALLED_EXTENSION_POINTS: typeof getActiveExtensionPoints;
     ENSURE_EXTENSION_POINTS_INSTALLED: typeof ensureInstalled;
 
@@ -97,8 +94,7 @@ export default function registerMessenger(): void {
     CLEAR_DYNAMIC_ELEMENTS: clearDynamicElements,
     UPDATE_DYNAMIC_ELEMENT: updateDynamicElement,
     RUN_EXTENSION_POINT_READER: runExtensionPointReader,
-    ENABLE_OVERLAY: enableOverlay,
-    DISABLE_OVERLAY: disableOverlay,
+
     INSTALLED_EXTENSION_POINTS: getActiveExtensionPoints,
     ENSURE_EXTENSION_POINTS_INSTALLED: ensureInstalled,
